+++ src/test/java/samples/training/src/training.java
@@ -1,0 +1,33 @@
+/*
+OC Volume - Java Speech Recognition Engine
+Copyright (C) 2002 OrangeCow organization
+
+This library is free software; you can redistribute it and/or
+modify it under the terms of the GNU Lesser General Public
+License as published by the Free Software Foundation; either
+version 2.1 of the License, or (at your option) any later version.
+
+This library is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public
+License along with this library; if not, write to the Free Software
+Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+
+Contact information:
+Please visit http://ocvolume.sourceforge.net.
+*/
+import org.oc.ocvolume.train;
+
+/**
+ * training program for OC Volume
+ * @author Danny Su
+ */
+public class training {
+    public static void main(String[] args) {
+        train t = new train();
+        t.show();
+    }
+}
